casting

+++ ivy/functional/backends/torch/experimental/manipulation.py
@@ -192,7 +192,7 @@ def take_along_axis(
             "arr and indices must have the same shape;"
             + f" got {arr.shape} vs {indices.shape}"        
         )
-    
+    indices = indices.long()
     return torch.take_along_dim(arr, indices, axis, out=out)
